Moved DatabaseException under the exceptions directory.
