+++ src/Libs/Exceptions/DatabaseException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Libs\Database;
+namespace App\Libs\Exceptions;
 
 use RuntimeException;
